+++ src/app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/Components/UnitView.tsx
@@ -1,7 +1,13 @@
 "use client";
-import { OakBox, OakGrid, OakGridArea } from "@oaknational/oak-components";
+import {
+  OakBox,
+  OakFlex,
+  OakGrid,
+  OakGridArea,
+} from "@oaknational/oak-components";
 
 import { LessonList } from "./LessonList";
+import { ProgrammeToggles } from "./ProgrammeToggles";
 
 import type { TeachersUnitOverviewData } from "@/node-lib/curriculum-api-2023/queries/teachersUnitOverview/teachersUnitOverview.schema";
 import SkipLink from "@/components/CurriculumComponents/OakComponentsKitchen/SkipLink";
@@ -19,6 +25,8 @@ export type UnitViewProps = Pick<
   | "lessons"
   | "unitIndex"
   | "unitCount"
+  | "tierOptionToggles"
+  | "subjectOptionToggles"
 >;
 
 export const UnitView = ({
@@ -33,6 +41,8 @@ export const UnitView = ({
   lessons,
   unitIndex,
   unitCount,
+  tierOptionToggles,
+  subjectOptionToggles,
 }: UnitViewProps) => {
   return (
     <OakBox $ph="spacing-40">
@@ -53,7 +63,22 @@ export const UnitView = ({
         >
           <SkipLink href="#lessons">Skip to lessons</SkipLink>
         </OakBox>
-        <OakGridArea $colSpan={[12, 5]} />
+        <OakGridArea $colSpan={[12, 5]}>
+          <OakFlex $flexDirection="column" $gap="spacing-32" $pb="spacing-56">
+            <ProgrammeToggles
+              heading="Learning tier (KS4)"
+              headingId="tier-toggle-heading"
+              unitSlug={unitSlug}
+              programmeToggles={tierOptionToggles}
+            />
+            <ProgrammeToggles
+              heading="Exam subject (KS4)"
+              headingId="subject-toggle-heading"
+              unitSlug={unitSlug}
+              programmeToggles={subjectOptionToggles}
+            />
+          </OakFlex>
+        </OakGridArea>
         <OakGridArea $colSpan={[12, 7]} $gap="spacing-56" id="lessons">
           <LessonList
             programmeSlug={programmeSlug}